refactor: split arch into struct-based domain files under rv32/
