+++ src/lib/panic.zig
@@ -1,8 +1,8 @@
 const console = @import("console");
-const sbi = @import("sbi");
+const arch = @import("arch");
 
 pub fn panic(comptime fmt: []const u8, args: anytype) noreturn {
     console.printf("\n[PANIC] " ++ fmt ++ "\n", args);
 
-    sbi.shutdown();
+    arch.Sbi.shutdown();
 }
